# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticação - Sistema de Música para DJ",
        "version": "1.0.0",
        "description": "API para registrar usuários, autenticar via JWT e acessar uma lista de músicas protegida."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/registro": {
            "post": {
                "summary": "Registrar novo usuário",
                "tags": [
                    "Usuário"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "usuario@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado."
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "Autenticar usuário e gerar token JWT",
                "tags": [
                    "Usuário"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "usuario@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token JWT gerado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas."
                    }
                }
            }
        },
        "/musicas": {
            "get": {
                "summary": "Obter lista de músicas (requer token JWT)",
                "tags": [
                    "Músicas"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de músicas retornada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "titulo": {
                                                "type": "string",
                                                "example": "Música A"
                                            },
                                            "artista": {
                                                "type": "string",
                                                "example": "DJ A"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token inválido."
                    },
                    "403": {
                        "description": "Token não fornecido."
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}